Move footer buttons to dedicated folder
