+++ app/assets/scripts/components/project/prime-panel/footer/prime-button.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import InfoButton from '../../common/info-button';
-import { ProjectMachineContext } from '../../../fsm/project';
-import selectors from '../../../fsm/project/selectors';
-import { SESSION_MODES } from '../../../fsm/project/constants';
-import * as guards from '../../../fsm/project/guards';
+import InfoButton from '../../../common/info-button';
+import { ProjectMachineContext } from '../../../../fsm/project';
+import selectors from '../../../../fsm/project/selectors';
+import { SESSION_MODES } from '../../../../fsm/project/constants';
+import * as guards from '../../../../fsm/project/guards';
 
 export function PrimeButton() {
   const actorRef = ProjectMachineContext.useActorRef();
