+++ app/assets/scripts/components/project/prime-panel/index.js
@@ -17,8 +17,8 @@ import TabbedBlock from '../../common/tabbed-block-body';
 import PredictTab from './tabs/predict';
 import RetrainTab from './tabs/retrain';
 import { UploadAoiModal } from './upload-aoi-modal';
-import { PrimeButton } from './prime-button';
-import { BatchPredictionPanel } from './batch-prediction-panel';
+import { PrimeButton } from './footer/prime-button';
+import { BatchPredictionPanel } from './footer/batch-prediction-panel';
 import { ProjectMachineContext } from '../../../fsm/project';
 import { SESSION_MODES } from '../../../fsm/project/constants';
 import selectors from '../../../fsm/project/selectors';
